refactor: rename to useGetIsolatedPoolsTreasuryTotals
